Make initTray injectable
